requirements.txt
cmake
ffmpeg
matplotlib==3.3.1
tensorboard==2.20.0
Unidecode==1.1.1
pysoundfile==0.9.0.post1
monotonic-align
g2pk2
ko_pron==1.3
jamo==0.4.1
# pyopenjtalk==0.2.0
jaconv
protobuf==3.19.0
Cython==0.29.21
librosa==0.8.0
numpy==1.23.5
scipy==1.10.1
